+++ src/IconButton.js
@@ -1,0 +1,26 @@
+import React, { PropTypes } from 'react';
+import icBEM from './utils/icBEM';
+
+import Button, { COMPONENT_NAME } from './Button';
+import IconLayout from './IconLayout';
+
+const rootClass = icBEM(COMPONENT_NAME)
+    .modifier('icon-only')
+    .toString({ stripBlock: true });
+
+function IconButton({ icon, ...otherProps }) {
+    return (
+        <Button className={rootClass} {...otherProps}>
+            <IconLayout icon={icon} />
+        </Button>
+    );
+}
+
+IconButton.propTypes = {
+    icon: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.element
+    ]).isRequired,
+};
+
+export default IconButton;